Rename QueryResultStatuses to MultiQueryResultStatus, change imports
