+++ src/app/common/components/QueryResultStatus/index.ts
@@ -1,0 +1,2 @@
+export * from './MultiQueryResultStatus';
+export * from './QueryResultStatus';
